requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
streamlit==1.27.2
pymongo==4.5.0
schedule==1.2.0
requests==2.31.0
Pillow==10.0.1
pandas==2.1.1
opencv-python==4.8.1.78
pydantic==2.4.2
python-multipart==0.0.6
